requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
line-bot-sdk>=3.11,<4
google-generativeai>=0.8
twstock>=1.3
python-dotenv>=1.0
